No flicking of the screen with clearScreen func
