+++ sources/move.c
@@ -7,41 +7,7 @@
 */
 
 #include "move.h"
-#include "fight.h"
 
-#include <unistd.h>
-#include <stdlib.h>
-
-#include <termios.h>
-#include <unistd.h>
-#include <fcntl.h>
-
-int kbhit(void)
-{
-    struct termios oldt, newt;
-    int ch;
-    int oldf;
-
-    tcgetattr(STDIN_FILENO, &oldt);
-    newt = oldt;
-    newt.c_lflag &= ~(ICANON | ECHO);
-    tcsetattr(STDIN_FILENO, TCSANOW, &newt);
-    oldf = fcntl(STDIN_FILENO, F_GETFL, 0);
-    fcntl(STDIN_FILENO, F_SETFL, oldf | O_NONBLOCK);
-
-    ch = getchar();
-
-    tcsetattr(STDIN_FILENO, TCSANOW, &oldt);
-    fcntl(STDIN_FILENO, F_SETFL, oldf);
-
-    if (ch != EOF)
-    {
-        ungetc(ch, stdin);
-        return 1;
-    }
-
-    return 0;
-}
 
 int isItem(Level *level, int h, int w)
 {
@@ -169,7 +135,7 @@ void reachCardinalPoint(Level *level)
 void showCurrentRoom(Level *level)
 {
     // clear screen
-    system("clear");
+    printf("\e[1;1H\e[2J");
 
     // color of the sceen
     system("echo '\e[032m'");
